fixed one bug in MSSQL manager
same bug was present in SQLite manager and while patched there, it wasnt
pathced for MSSQL. also added the Logger class for logging exceptions
and events. this will require the SQLite JDBC driver be included in the
redistributable regardless of if the program users intend to use SQLite
or not. formatted source